requirements.txt
streamlit
langgraph
groq
trafilatura
newspaper3k
playwright
curl_cffi
langfuse
playwright-stealth
python-dotenv
pyyaml
google-generativeai
